Export VideoAtomType
